Normalizer
Added SimpleStats.h to Normalizer.cpp
